backend/requirements.txt
Flask
newsapi-python
requests
python-dotenv
openai       # Optional, if using OpenAI for content synthesis
